MVP time
Added the distribution function so we can have something to show for it. Running distribute destroys the playerList but does not destroy the captainList, so if something goes wrong people can re-join and an admin can hit !distribute again to see where it goes. In testing, this works. Yes, I know it's messy.